Hydra: extract multi_array support to own file
