+++ include/nix/Hydra.hpp
@@ -15,8 +15,6 @@
 #include <type_traits>
 #include <valarray>
 
-#include <boost/multi_array.hpp>
-
 #ifndef NIX_HYDRA_H
 #define NIX_HYDRA_H
 
@@ -235,52 +233,6 @@ public:
             return;
 
         value.resize(dims[0]);
-    }
-};
-
-
-template<typename T, size_t N>
-class data_traits<boost::multi_array<T, N> > {
-public:
-
-    typedef boost::multi_array<T, N> value_type;
-    typedef value_type&              reference;
-    typedef const value_type&        const_reference;
-
-    typedef T        element_type;
-    typedef T*       element_pointer;
-    typedef const T* const_element_pointer;
-
-    static DataType data_type(const_reference val) {
-        return to_data_type<element_type>::value;
-    }
-
-    static NDSize shape(const value_type &value) {
-        NDSize ndsize(N);
-        const size_t *ma_shape = value.shape();
-        std::transform(ma_shape, ma_shape + N, ndsize.data(),
-                       [](NDSize::const_reference val) { return static_cast<T>(val);});
-        return ndsize;
-    }
-
-    static size_t num_elements(const_reference value) {
-        return  value.num_elements();;
-    }
-
-    static const_element_pointer get_data(const value_type& value) {
-        return value.data();
-    }
-
-    static element_pointer get_data(value_type &value) {
-        return value.data();
-    }
-
-    static void resize(reference value, const NDSize &dims) {
-        if (dims.size() != N) {
-            throw InvalidRank("Cannot change rank of multiarray");
-        }
-
-        value.resize(dims);
     }
 };
