+++ test/TestDataSet.hpp
@@ -6,10 +6,10 @@
 // modification, are permitted under the terms of the BSD License. See
 // LICENSE file in the root of the Project.
 
-#include <nix/NDArray.hpp>
+#include <nix/hydra/multiArray.hpp>
 #include <nix.hpp>
+
 #include <nix/hdf5/Group.hpp>
-#include <nix/hdf5/DataSetHDF5.hpp>
 
 #include <iostream>
 #include <sstream>
