+++ test/TestGroup.hpp
@@ -8,7 +8,9 @@
 //
 // Author: Christian Kellner <kellner@bio.lmu.de>
 
+#include <nix/hydra/multiArray.hpp>
 #include <nix.hpp>
+
 #include <nix/hdf5/Group.hpp>
 
 #include <iostream>
