+++ test/TestMultiTag.hpp
@@ -6,8 +6,9 @@
 // modification, are permitted under the terms of the BSD License. See
 // LICENSE file in the root of the Project.
 
-
+#include <nix/hydra/multiArray.hpp>
 #include <nix.hpp>
+
 #include <nix/hdf5/MultiTagHDF5.hpp>
 #include <nix/util/util.hpp>
 #include <nix/util/dataAccess.hpp>
