+++ test/TestTag.hpp
@@ -6,7 +6,9 @@
 // modification, are permitted under the terms of the BSD License. See
 // LICENSE file in the root of the Project.
 
+#include <nix/hydra/multiArray.hpp>
 #include <nix.hpp>
+
 #include <nix/hdf5/TagHDF5.hpp>
 #include <nix/util/util.hpp>
