+++ test/TestValidate.hpp
@@ -6,6 +6,7 @@
 // modification, are permitted under the terms of the BSD License. See
 // LICENSE file in the root of the Project.
 
+#include <nix/hydra/multiArray.hpp>
 #include <nix.hpp>
 
 #include <iostream>
